// src/services/api/models/livestream.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Category } from './category';
import { LivestreamStatus } from './livestream-status';
import { LivestreamType } from './livestream-type';
/**
 * 
 * @export
 * @interface Livestream
 */
export interface Livestream {
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof Livestream
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof Livestream
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof Livestream
     */
    updated?: Date;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    name: string;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    timezone?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof Livestream
     */
    startTime: Date;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    hostUserId: string;
    /**
     * 
     * @type {LivestreamStatus}
     * @memberof Livestream
     */
    status?: LivestreamStatus;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    imageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    promoImageUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    promoDescription?: string | null;
    /**
     * 
     * @type {string}
     * @memberof Livestream
     */
    mediaUrl?: string | null;
    /**
     * 
     * @type {LivestreamType}
     * @memberof Livestream
     */
    livestreamType?: LivestreamType;
    /**
     * 
     * @type {boolean}
     * @memberof Livestream
     */
    isNotified?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof Livestream
     */
    lastLiveUpdated?: Date | null;
    /**
     * 
     * @type {number}
     * @memberof Livestream
     */
    broadcasterCount?: number;
    /**
     * 
     * @type {number}
     * @memberof Livestream
     */
    viewerCount?: number;
    /**
     * 
     * @type {Array<Category>}
     * @memberof Livestream
     */
    categories?: Array<Category> | null;
}