// src/services/api/models/social-post.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Conifer.NFT.API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: 1.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { SocialPostAttributes } from './social-post-attributes';
import { SocialPostType } from './social-post-type';
import { UserRole } from './user-role';
/**
 * 
 * @export
 * @interface SocialPost
 */
export interface SocialPost {
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    id?: string;
    /**
     * 
     * @type {boolean}
     * @memberof SocialPost
     */
    isDeleted?: boolean;
    /**
     * 
     * @type {Date}
     * @memberof SocialPost
     */
    created?: Date;
    /**
     * 
     * @type {Date}
     * @memberof SocialPost
     */
    updated?: Date;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    userId?: string;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    post?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    parentPostId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    categoryId?: string;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    communityId?: string;
    /**
     * 
     * @type {SocialPostType}
     * @memberof SocialPost
     */
    postType?: SocialPostType;
    /**
     * 
     * @type {SocialPostAttributes}
     * @memberof SocialPost
     */
    socialPostAttributes?: SocialPostAttributes;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    userName?: string | null;
    /**
     * 
     * @type {UserRole}
     * @memberof SocialPost
     */
    userRole?: UserRole;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    profilePicUrl?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SocialPost
     */
    coverImageUrl?: string | null;
    /**
     * 
     * @type {boolean}
     * @memberof SocialPost
     */
    isProfilePicVideo?: boolean;
}